--- Assets/Scripts/Model.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 99a575e60ac5fc34bb11526bb58cdd91
-folderAsset: yes
-timeCreated: 1498385237
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
