--- Assets/Scripts/Orchestrator.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: f863f2511b546b94fa9f0852bbd1f41b
-folderAsset: yes
-timeCreated: 1498385237
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
